requirements.txt
requests
Wikipedia
webbrowser
pywhatkit
pyttsx3
SpeechRecognition
python-dotenv
pyaudio
opencv-python
python-decouple
smtplib
